Remove src/libsyntax/parse/attr.rs
This file is no longer needed as attribute parsing is no longer part of a trait.
