+++ src/libsyntax/parse/mod.rs
@@ -32,7 +32,6 @@ pub mod parser;
 
 pub mod lexer;
 pub mod token;
-pub mod attr;
 
 pub mod common;
 pub mod classify;
